Worked on trash can
I’ve got a mesh collider set for the trash can in addition with the capsule collider, which is positioned such that you can only collide with it from the top. My goal is to make a trigger that happens when you hit the capsule collider but not the mesh collider that will hide the player until you press space. I have attempted to do that with the HideOnCollide script, but it has not worked so far.