Updated Henrys Patrol.
Changed how henrys patrol works now added a waypoint system under paths gameobject. you can add waypoints and he will go to it in the order the are placed in the hierarchy. you can attach a path by attaching the overall path and putting it on the serialized variable . right now it is only on henry but later it should be decoupled so that any character can use it